+++ includes/Observers/AttachmentUrlObserver.php
@@ -18,7 +18,7 @@ class AttachmentUrlObserver implements ObserverInterface {
     /**
      * Constructor.
      *
-     * @param S3_Provider $cloudProvider
+     * @param S3_Provider $cloudProvider The cloud storage provider instance.
      */
     public function __construct( S3_Provider $cloudProvider ) {
         $this->cloudProvider = $cloudProvider;