+++ includes/Traits/OffloaderTrait.php
@@ -38,9 +38,9 @@ trait OffloaderTrait {
 
         // Generate a new version
         if ( ! advmo_is_media_organized_by_year_month() ) {
-            $new_version = date( 'YmdHis' );
+            $new_version = gmdate( 'YmdHis' );
         } else {
-            $new_version = date( 'dHis' );
+            $new_version = gmdate( 'dHis' );
         }
 
         // Save the new version in post meta